Rename bluegreen_deploy -> zdd.
There's a symlink from bluegreen_deploy.py -> zdd.py to prevent breakage
on existing deployments.
