+++ tests/test_zdd.py
@@ -1,5 +1,5 @@
 import unittest
-import bluegreen_deploy
+import zdd
 import mock
 import json
 import time
@@ -31,12 +31,12 @@ class MyResponse:
 
 def _load_listeners():
     with open('tests/haproxy_stats.csv') as f:
-        return bluegreen_deploy.parse_haproxy_stats(f.read())
+        return zdd.parse_haproxy_stats(f.read())
 
 
 class TestBluegreenDeploy(unittest.TestCase):
 
-    @mock.patch('bluegreen_deploy.scale_marathon_app_instances')
+    @mock.patch('zdd.scale_marathon_app_instances')
     def test_scale_new_app_instances_up_50_percent(self, mock):
         """When scaling new_app instances, increase instances by 50% of
            existing instances if we have not yet met or surpassed the amount
@@ -51,11 +51,11 @@ class TestBluegreenDeploy(unittest.TestCase):
         old_app = {'instances': 30}
         args = Arguments()
         args.initial_instances = 5
-        bluegreen_deploy.scale_new_app_instances(args, new_app, old_app)
+        zdd.scale_new_app_instances(args, new_app, old_app)
         mock.assert_called_with(
           args, new_app, 15)
 
-    @mock.patch('bluegreen_deploy.scale_marathon_app_instances')
+    @mock.patch('zdd.scale_marathon_app_instances')
     def test_scale_new_app_instances_to_target(self, mock):
         """When scaling new instances up, if we have met or surpassed the
            amount of instances deployed for old_app, go right to our
@@ -70,20 +70,20 @@ class TestBluegreenDeploy(unittest.TestCase):
         old_app = {'instances': 8}
         args = Arguments()
         args.initial_instances = 5
-        bluegreen_deploy.scale_new_app_instances(args, new_app, old_app)
+        zdd.scale_new_app_instances(args, new_app, old_app)
         mock.assert_called_with(
           args, new_app, 30)
 
     def test_find_drained_task_ids(self):
         listeners = _load_listeners()
         haproxy_instance_count = 2
-        apps = json.loads(open('tests/bluegreen_app_blue.json').read())
+        apps = json.loads(open('tests/zdd_app_blue.json').read())
         app = apps['apps'][0]
 
         results = \
-            bluegreen_deploy.find_drained_task_ids(app,
-                                                   listeners,
-                                                   haproxy_instance_count)
+            zdd.find_drained_task_ids(app,
+                                      listeners,
+                                      haproxy_instance_count)
 
         assert app['tasks'][0]['id'] in results  # 2 l's down, no sessions
         assert app['tasks'][1]['id'] not in results  # 1 l up, 1 down
@@ -92,30 +92,30 @@ class TestBluegreenDeploy(unittest.TestCase):
     def test_find_draining_task_ids(self):
         listeners = _load_listeners()
         haproxy_instance_count = 2
-        apps = json.loads(open('tests/bluegreen_app_blue.json').read())
+        apps = json.loads(open('tests/zdd_app_blue.json').read())
         app = apps['apps'][0]
 
         results = \
-            bluegreen_deploy.find_draining_task_ids(app,
-                                                    listeners,
-                                                    haproxy_instance_count)
+            zdd.find_draining_task_ids(app,
+                                       listeners,
+                                       haproxy_instance_count)
 
         assert app['tasks'][0]['id'] in results  # 2 l's down, no sessions
         assert app['tasks'][1]['id'] not in results  # 1 l up, 1 down
         assert app['tasks'][2]['id'] in results  # 2 l's down, 1 w/ scur/qcur
 
     def test_get_svnames_from_tasks(self):
-        apps = json.loads(open('tests/bluegreen_app_blue.json').read())
+        apps = json.loads(open('tests/zdd_app_blue.json').read())
         tasks = apps['apps'][0]['tasks']
 
-        task_svnames = bluegreen_deploy.get_svnames_from_tasks(tasks)
+        task_svnames = zdd.get_svnames_from_tasks(tasks)
 
         assert '10_0_6_25_16916' in task_svnames
         assert '10_0_6_25_31184' in task_svnames
 
     def test_parse_haproxy_stats(self):
         with open('tests/haproxy_stats.csv') as f:
-            results = bluegreen_deploy.parse_haproxy_stats(f.read())
+            results = zdd.parse_haproxy_stats(f.read())
 
             assert results[1].pxname == 'http-in'
             assert results[1].svname == 'IPv4-direct'
@@ -128,18 +128,25 @@ class TestBluegreenDeploy(unittest.TestCase):
         # TODO(BM): This test is naive. An end-to-end test would be nice.
         args = Arguments()
         args.pre_kill_hook = 'myhook'
-        app = {
-            'id': 'myApp'
+        old_app = {
+            'id': 'oldApp'
+        }
+        new_app = {
+            'id': 'newApp'
         }
         tasks_to_kill = ['task1', 'task2']
 
-        bluegreen_deploy.execute_pre_kill_hook(args, app, tasks_to_kill)
+        zdd.execute_pre_kill_hook(args,
+                                  old_app,
+                                  tasks_to_kill,
+                                  new_app)
 
         mock.assert_called_with([args.pre_kill_hook,
-                                 '{"id": "myApp"}',
-                                 '["task1", "task2"]'])
+                                 '{"id": "oldApp"}',
+                                 '["task1", "task2"]',
+                                 '{"id": "newApp"}'])
 
-    @mock.patch('bluegreen_deploy.fetch_combined_haproxy_stats',
+    @mock.patch('zdd.fetch_combined_haproxy_stats',
                 mock.Mock(side_effect=lambda _: _load_listeners()))
     def test_fetch_app_listeners(self):
         app = {
@@ -149,7 +156,7 @@ class TestBluegreenDeploy(unittest.TestCase):
                 }
               }
 
-        app_listeners = bluegreen_deploy.fetch_app_listeners(app, [])
+        app_listeners = zdd.fetch_app_listeners(app, [])
 
         assert app_listeners[0].pxname == 'foobar_8080'
         assert len(app_listeners) == 1
@@ -158,7 +165,7 @@ class TestBluegreenDeploy(unittest.TestCase):
                 mock.Mock(side_effect=lambda hostname:
                           (hostname, [], ['127.0.0.1', '127.0.0.2'])))
     def test_get_marathon_lb_urls(self):
-        marathon_lb_urls = bluegreen_deploy.get_marathon_lb_urls(Arguments())
+        marathon_lb_urls = zdd.get_marathon_lb_urls(Arguments())
 
         assert 'http://127.0.0.1:9090' in marathon_lb_urls
         assert 'http://127.0.0.2:9090' in marathon_lb_urls
@@ -166,14 +173,14 @@ class TestBluegreenDeploy(unittest.TestCase):
 
     @mock.patch('requests.get',
                 mock.Mock(side_effect=lambda k, auth:
-                          MyResponse('tests/bluegreen_app_blue.json')))
+                          MyResponse('tests/zdd_app_blue.json')))
     def test_simple(self):
         # This test just checks the output of the program against
         # some expected output
         from six import StringIO
 
         out = StringIO()
-        bluegreen_deploy.do_bluegreen_deploy(Arguments(), out)
+        zdd.do_zdd(Arguments(), out)
         output = json.loads(out.getvalue())
         output['labels']['HAPROXY_DEPLOYMENT_STARTED_AT'] = ""
